Создал отдельный класс smart и вынес в него компоненты для перерисовки. Исправил ошибку с обработкой выбора опций (цена опции записывалась в значение цены всей точки)
